refactor baichuan1 7b and 13b
